Refactor MQTT classes
- Removed connection code from mqtt_handler

- Added extra error handling and logging to mqtt_connector
